requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.32.2
langchain>=0.2.0
langchain-groq==0.1.3
langgraph>=0.2.14
pandas==2.2.1
scikit-learn==1.4.1.post1
matplotlib==3.8.3
seaborn==0.13.2
python-dotenv==1.0.1
pydantic==2.6.4
requests==2.31.0
